+++ packages/terra-application-header-layout/tests/wdio/application-header-spec.js
@@ -13,12 +13,10 @@ describe('Application Header Layout', () => {
 
   describe('Displays a application header layout with themeable styles', () => {
     beforeEach(() => {
+      browser.setViewportSize(Terra.viewports('huge')[0]);
       browser.url('/#/raw/tests/terra-application-header-layout/application-header-layout/application-header-default');
       browser.waitForVisible('#test-header');
     });
-
-    Terra.should.matchScreenshot({ viewports, selector: '#test-header' });
-    Terra.should.beAccessible({ viewports, context: '#test-header' });
 
     Terra.should.themeEachCustomProperty('#test-header', {
       '--terra-application-header-layout-logo-min-width': '5rem',