+++ packages/terra-disclosure-manager/src/DisclosureManagerContext.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const DisclosureManagerContext = React.createContext();
+
+export default DisclosureManagerContext;