+++ packages/terra-application-name/CHANGELOG.md
@@ -4,6 +4,20 @@ Changelog
 Unreleased
 ----------
 
+2.22.0 - (December 5, 2018)
+------------------
+### Changed
+* Minor dependency update
+* Update example to use terra-framework's checked in image
+
+2.21.0 - (December 3, 2018)
+------------------
+### Changed
+* Wdio tests changed to use themeCombinationOfCustomProperties
+
+### Removed
+* Removed dependency on terra-app-delegate package.
+
 2.20.0 - (November 15, 2018)
 ------------------
 ### Changed