+++ packages/terra-application-utility/CHANGELOG.md
@@ -4,6 +4,19 @@ Changelog
 Unreleased
 ----------
 
+1.24.0 - (December 5, 2018)
+------------------
+### Changed
+* Minor dependency update
+
+1.23.0 - (December 3, 2018)
+------------------
+### Changed
+* Wdio tests changed to use themeCombinationOfCustomProperties
+
+### Removed
+* Removed dependency on terra-app-delegate package.
+
 1.22.0 - (November 15, 2018)
 ------------------
 ### Changed