+++ packages/terra-disclosure-manager/CHANGELOG.md
@@ -4,6 +4,21 @@ ChangeLog
 Unreleased
 ----------
 
+3.1.0 - (December 5, 2018)
+------------------
+### Changed
+* Minor dependency update
+
+3.0.0 - (December 3, 2018)
+------------------
+### Breaking Change
+* Added Context-based API for consumer-facing DisclosureManager APIs
+* Removed injection of `app` prop into child/disclosure components.
+
+### Added
+* Added exports for withDisclosureManager and disclosureManagerShape
+* Added Upgrade Guide for conversion from v2.x to v3.x
+
 2.15.0 - (November 15, 2018)
 ------------------
 ### Changed