+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -4,6 +4,19 @@ Changelog
 Unreleased
 ----------
 
+3.1.0 - (December 5, 2018)
+------------------
+### Changed
+* Minor dependency update
+
+3.0.0 - (December 3, 2018)
+------------------
+### Breaking Change
+* Removed instances of the NavigationLayout and related components injecting an `app` prop into child components.
+
+### Removed
+* Removed dependency on terra-app-delegate package.
+
 2.22.0 - (November 15, 2018)
 ------------------
 ### Changed