+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+1.18.0 - (December 5, 2018)
+------------------
+### Changed
+* Minor dependency update
+
+1.17.0 - (December 3, 2018)
+------------------
+### Changed
+* Wdio tests changed to use themeCombinationOfCustomProperties
+
 1.16.0 - (November 15, 2018)
 ------------------
 ### Changed