+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -3,8 +3,23 @@ Changelog
 
 Unreleased
 ----------
+
+1.15.0 - (December 5, 2018)
+------------------
+### Changed
+* Minor dependency update
+
+1.14.0 - (December 3, 2018)
+------------------
+### Added
+* Added missing ar, es-ES, es-US, fr-FR, and pt-BR translations.
+
 ### Changed
 * Minor improvements to doc examples
+* Updated examples to consume terra-disclosure-manager v3.x
+
+### Removed
+* Removed dependency on terra-app-delegate package.
 
 1.13.0 - (November 15, 2018)
 ------------------