+++ packages/terra-slide-panel/CHANGELOG.md
@@ -4,6 +4,19 @@ Changelog
 Unreleased
 ----------
 
+2.23.0 - (December 5, 2018)
+------------------
+### Changed
+* Minor dependency update
+
+2.22.0 - (December 3, 2018)
+------------------
+### Changed
+* Wdio tests changed to use themeCombinationOfCustomProperties
+
+### Fixed
+* Slide panel's aria-hidden value when the panel is toggled
+
 2.21.0 - (November 15, 2018)
 ------------------
 ### Changed