+++ packages/terra-theme-provider/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+2.19.0 - (December 5, 2018)
+------------------
+### Changed
+* Minor dependency update
+
+2.18.0 - (December 3, 2018)
+------------------
+### Changed
+* Minor dependency bump
+
 2.17.0 - (November 15, 2018)
 ------------------
 ### Changed